pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "afterburner-customer-bot"
version = "0.1.0"
description = "Conversational AI customer discovery bot with personality framework"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
afterburner-bot = "cli:main"

[tool.setuptools.packages.find]
include = ["bot*", "evaluations*"]
